third_party/blink/web_tests/virtual/mutation-events-enabled/dom/mutation-event-tests/fast/forms/mutation-event-recalc-expected.txt

CONSOLE WARNING: Listener added for a 'DOMNodeInserted' mutation event. This event type is no longer supported, and will be removed from this browser VERY soon. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
CONSOLE WARNING: Listener added for a 'DOMNodeInsertedIntoDocument' mutation event. This event type is no longer supported, and will be removed from this browser VERY soon. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
CONSOLE WARNING: Listener added for a 'DOMSubtreeModified' mutation event. This event type is no longer supported, and will be removed from this browser VERY soon. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
Test that HTMLSelectElement DOM is in a consistent state when handling mutation events.

SUCCESS
